+++ app/src/main/kotlin/me/xizzhu/android/joshua/core/repository/remote/RemoteTranslationService.kt
@@ -39,5 +39,5 @@ data class RemoteTranslation(
 interface RemoteTranslationService {
     suspend fun fetchTranslations(): List<RemoteTranslationInfo>
 
-    suspend fun fetchTranslation(channel: SendChannel<Int>, translationInfo: RemoteTranslationInfo): RemoteTranslation
+    suspend fun fetchTranslation(downloadProgress: SendChannel<Int>, translationInfo: RemoteTranslationInfo): RemoteTranslation
 }